# src/main/resources/docs/api.yaml
openapi: 3.1.0
info:
  title: HFF Data Studio API
  description: HFF Data Intelligence Platform - CATalyst API
  version: 1.0.0
  contact:
    name: Data Analyst Service - HFF
    email: data-analyst@hff.re
  license:
    name: HFF Internal Use - Restricted Distribution
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://catalyst.hff.re/api/v1
    description: Production server
...
paths:
...
  /admin/users/category/{categoryId}:
    get:
      tags: [Users]
      summary: Get users by category
      operationId: getUsersByCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageUserResponse"
...
  /admin/categories:
    get:
      tags: [Categories]
      summary: Get all categories
      operationId: getAllCategories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserCategoryResponse"
    post:
      tags: [Categories]
      summary: Create category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreationRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCategoryResponse"
  /admin/categories/{categoryId}:
    get:
      tags: [Categories]
      summary: Get category by ID
      operationId: getCategoryById
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCategoryResponse"
    put:
      tags: [Categories]
      summary: Update category
      operationId: updateCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreationRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCategoryResponse"
    delete:
      tags: [Categories]
      summary: Delete category
      operationId: deleteCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Category deleted
  /admin/categories/{categoryId}/permissions:
    get:
      tags: [Categories]
      summary: Get category permissions
      operationId: getCategoryPermissions
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionResponse"
    put:
      tags: [Categories]
      summary: Update category permissions
      operationId: updateCategoryPermissions
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionUpdateRequest"
      responses:
        "200":
          description: Permissions updated
...
  /datasets/{datasetId}/exploration/report:
    get:
      tags: [Exploration]
      summary: Get exploration report
      operationId: getExplorationReport
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Exploration report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplorationReportResponse"
    post:
      tags: [Exploration]
      summary: Generate exploration report
      operationId: generateExplorationReport
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Exploration report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplorationReportResponse"
  /datasets/{datasetId}/exploration/report/export:
    get:
      tags: [Exploration]
      summary: Export exploration report
      operationId: exportExplorationReport
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          schema:
            type: string
            enum: [PDF, HTML]
            default: PDF
      responses:
        "200":
          description: Report file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
  /datasets/{datasetId}/exploration/summary:
    get:
      tags: [Exploration]
      summary: Get exploration summary
      operationId: getExplorationSummary
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Exploration summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplorationReportResponse"

  # ==================== CLEANING ====================
...
  /exports:
    post:
      tags: [Exports]
      summary: Export results
      operationId: exportResults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "201":
          description: Export created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
    get:
      tags: [Exports]
      summary: Get exports
      operationId: getExports
      parameters:
        - name: format
          in: query
          schema:
            $ref: "#/components/schemas/ExportFormat"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of exports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageExportResponse"
  /exports/{exportId}:
    get:
      tags: [Exports]
      summary: Get export by ID
      operationId: getExportById
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Export details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
    delete:
      tags: [Exports]
      summary: Delete export
      operationId: deleteExport
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Export deleted
  /exports/{exportId}/download:
    get:
      tags: [Exports]
      summary: Download export
      operationId: downloadExport
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Export file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /exports/execution/{executionId}:
    get:
      tags: [Exports]
      summary: Get exports by execution
      operationId: getExportsByExecution
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of exports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageExportResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        lastName:
          type: string
        firstName:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/UserCategoryResponse"
...
    UserCategoryResponse:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        label:
          type: string
        description:
          type: string
        accessLevel:
          type: integer
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionResponse"
    PermissionResponse:
      type: object
      properties:
        permissionId:
          type: string
          format: uuid
        code:
          type: string
        label:
          type: string
        module:
          type: string
    CategoryCreationRequest:
      type: object
      properties:
        label:
          type: string
        description:
          type: string
        accessLevel:
          type: integer
      required: [label, accessLevel]
    PermissionUpdateRequest:
      type: object
      properties:
        permissionIds:
          type: array
          items:
            type: string
            format: uuid
...
    ExplorationReportResponse:
      type: object
      properties:
        reportId:
          type: string
          format: uuid
        generatedAt:
          type: string
          format: date-time
        totalRows:
          type: integer
        duplicateCount:
          type: integer
        missingValuesCount:
          type: integer
        qualityScore:
          type: number
        reportPdfPath:
          type: string
        datasetId:
          type: string
          format: uuid
        datasetName:
          type: string
        columnStatistics:
          type: object

    # ===== CLEANING =====
...
    ExportFormat:
      type: string
      enum: [CSV, XLSX, PNG, ZIP]
    ExportRequest:
      type: object
      properties:
        executionId:
          type: string
          format: uuid
        exportFormat:
          $ref: "#/components/schemas/ExportFormat"
        resultIds:
          type: array
          items:
            type: string
            format: uuid
      required: [executionId, exportFormat]
    ExportResponse:
      type: object
      properties:
        exportId:
          type: string
          format: uuid
        exportFormat:
          $ref: "#/components/schemas/ExportFormat"
        formatLabel:
          type: string
        exportedAt:
          type: string
          format: date-time
        filePath:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        fileSizeFormatted:
          type: string
        executionId:
          type: string
          format: uuid
        executionStatus:
          type: string
        userId:
          type: string
          format: uuid
        userFullName:
          type: string
        downloadUrl:
          type: string

    # ===== AUDIT =====
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        isFirst:
          type: boolean
        isLast:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    PageUserResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/UserResponse"
...
    PageExportResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/ExportResponse"
...
security:
  - bearerAuth: []